refactor: extract api provider logic
